refactor: #63 구인글 입력 공간 분리
