+++ src/pages/create/recruitCreatePage/RecruitCreatePage.styled.ts
@@ -9,220 +9,6 @@ const RecruitCreatePage = styled.div`
 		margin-top: 1.6rem;
 	}
 
-	.wrapper {
-		width: 100%;
-		margin-top: 2rem;
-		padding-bottom: 0.5rem;
-		border-radius: 0.5rem;
-
-		.container {
-			width: 100%;
-			margin: 0 auto;
-			display: flex;
-			flex-direction: column;
-			gap: 5rem;
-		}
-
-		.container__info {
-		}
-
-		.container__tag {
-		}
-
-		.container__img {
-			margin-top: 7rem;
-
-			.container__img-input {
-				padding: 10px 0px;
-
-				input {
-					display: none;
-				}
-
-				label {
-					display: block;
-					width: 100%;
-					height: 4.875rem;
-					flex-shrink: 0;
-					border-radius: 0.75rem;
-					border: 0.75px solid #e3e3e3;
-					background: #fff;
-					outline: none;
-					padding-left: 2.1rem;
-					box-sizing: border-box;
-					color: #a7a7a7;
-					font-size: 1.3rem;
-					font-style: normal;
-					font-weight: 400;
-					line-height: 1.8rem; /* 90% */
-					letter-spacing: 0.02rem;
-					display: flex;
-					align-items: center;
-				}
-
-				.haveFile {
-					color: #000;
-				}
-			}
-		}
-
-		.container__member {
-			position: relative;
-			margin-top: 4.35rem;
-			width: 100%;
-
-			.container__member-title {
-				display: flex;
-				justify-content: space-between;
-
-				button {
-					display: flex;
-					width: 10rem;
-					height: 3.5rem;
-					padding: 1rem;
-					justify-content: center;
-					align-items: center;
-					gap: 1rem;
-					border: none;
-					outline: none;
-					border-radius: 0.4rem;
-					background: var(--main-color, #5877fc);
-					color: #fff;
-				}
-			}
-
-			.container__member-area {
-				min-height: 24rem;
-				width: 100%;
-				background-color: #fff;
-				border: 1px solid #dcdcdc;
-				border-radius: 1rem;
-				padding: 3rem 4rem;
-				display: grid;
-				grid-template-columns: 1fr 1fr 1fr;
-				grid-template-rows: 1fr 1fr;
-				gap: 1rem 1rem;
-			}
-
-			.container__member-area__element {
-				width: 24rem;
-				height: 14rem;
-				flex-shrink: 0;
-				border-radius: 1rem;
-				border: 1px dashed #cdcdcd;
-				background: #fdfdfd;
-			}
-
-			.member {
-				position: relative;
-				width: 24rem;
-
-				.delete {
-					position: absolute;
-					top: 1rem;
-					right: 1.2rem;
-					font-size: 1.5rem;
-					cursor: pointer;
-				}
-			}
-
-			.controll {
-				/* display: flex; */
-				gap: 0.5rem;
-				margin-bottom: 30px;
-
-				button {
-					border: none;
-					width: 5rem;
-					height: 1.5rem;
-					border-radius: 5px;
-					cursor: pointer;
-				}
-
-				button:nth-child(2) {
-					background-color: #00b894;
-					color: #fff;
-				}
-
-				button:nth-child(3) {
-					background-color: #e17055;
-					color: #fff;
-				}
-			}
-
-			.container__member-add {
-				display: flex;
-				justify-content: center;
-				align-items: center;
-
-				.addition {
-					width: 7rem;
-					height: 7rem;
-					flex-shrink: 0;
-					display: flex;
-					justify-content: center;
-					align-items: center;
-					border-radius: 50%;
-					color: #000;
-					cursor: pointer;
-					opacity: 0.8;
-					filter: drop-shadow(0px 3px 18.75px rgba(0, 0, 0, 0.11));
-
-					.addition-container {
-					}
-				}
-			}
-		}
-	}
-
-	.wrapper-recruit {
-		.title {
-			display: flex;
-			flex-direction: column;
-			margin-top: 4rem;
-
-			input {
-				width: 48%;
-				height: 4.875rem;
-				flex-shrink: 0;
-				border-radius: 0.75rem;
-				border: 0.1rem solid #e3e3e3;
-				background: #fff;
-				outline: none;
-				padding-left: 2.1rem;
-				box-sizing: border-box;
-				margin-top: 1.2rem;
-			}
-		}
-
-		.container__intro {
-			margin-top: 4rem;
-			div:nth-child(1) {
-				margin-top: 1.2rem;
-				.editor {
-					background-color: #fff;
-					border-radius: 0.1rem;
-				}
-			}
-
-			.ql-container.ql-snow {
-				border: 0.1rem solid #e3e3e3;
-				border-bottom-right-radius: 0.75rem;
-				border-bottom-left-radius: 0.75rem;
-				height: 30.825rem;
-			}
-
-			.ql-toolbar.ql-snow {
-				border: 0.75px solid #e3e3e3;
-				border-top-right-radius: 0.75rem;
-				border-top-left-radius: 0.75rem;
-				box-sizing: border-box;
-				font-family: 'Helvetica Neue', 'Helvetica', 'Arial', sans-serif;
-				padding: 0.8rem;
-			}
-		}
-	}
-
 	.container__controller {
 		margin-top: 2.77rem;
 		display: flex;
