+++ src/pages/create/recruitCreatePage/wrappers/RecruitInfoWrapper.styled.ts
@@ -1,0 +1,26 @@
+import styled from 'styled-components';
+
+const RecruitInfoWrapper = styled.section`
+	width: 100%;
+	margin-top: 2rem;
+	padding-bottom: 0.5rem;
+	border-radius: 0.5rem;
+
+	.container {
+		width: 100%;
+		margin: 0 auto;
+		display: flex;
+		flex-direction: column;
+		gap: 5rem;
+	}
+
+	.container__info {
+	}
+
+	.container__tag {
+	}
+`;
+
+const S = { RecruitInfoWrapper };
+
+export default S;
