+++ src/pages/create/recruitCreatePage/wrappers/RecruitInfoWrapper.tsx
@@ -1,0 +1,24 @@
+import React from 'react';
+import { InputContainer, Subtitle, MeeteamTag } from '../../../../components';
+import S from './RecruitInfoWrapper.styled';
+
+const RecruitInfoWrapper = () => {
+	return (
+		<S.RecruitInfoWrapper>
+			<section className='container'>
+				<article className='container__info'>
+					<InputContainer />
+				</article>
+				<article className='container__tag'>
+					<Subtitle>{'태그'}</Subtitle>
+					<MeeteamTag />
+				</article>
+				<article className='container__role'>
+					<Subtitle>{'역할'}</Subtitle>
+				</article>
+			</section>
+		</S.RecruitInfoWrapper>
+	);
+};
+
+export default RecruitInfoWrapper;
