+++ src/types/index.ts
@@ -1,6 +1,6 @@
 import type { SignUpPayload } from './payload';
 import type { UserReponse, User } from './response';
 import type { CustomInstance } from './api';
-import type { TitleInfo } from './information';
+import type { TitleInfo, OptionList } from './information';
 
-export { SignUpPayload, UserReponse, User, CustomInstance, TitleInfo };
+export { SignUpPayload, UserReponse, User, CustomInstance, TitleInfo, OptionList };
