+++ src/types/information.ts
@@ -2,3 +2,7 @@ export interface TitleInfo {
 	title?: string;
 	descriptions?: string[];
 }
+
+export interface OptionList {
+	options: string[];
+}
